Errorhandling - or better the display of it - improved (Teksura)
Double != float in WINE - Entitymodel updated
PostSINner now allowes NPCAdmins to save files, even if they are not in the UserRightsList of the character
UploadSinner now does the same

+++ ChummerHub/Models/V1/Tag.cs
@@ -26,7 +26,7 @@ namespace ChummerHub.Models.V1
         public string TagValue { get; set; }
 
 
-        public double? TagValueDouble { get; set; }
+        public float? TagValueDouble { get; set; }
 
         /// <summary>
         /// This has NO FUNCTION and is only here for Debugging reasons.
